fix: email sending was not linked in new modal + removed offline from readme
